adding experimental API object to PCA
